@@ -1,15 +1,18 @@
 [package]
 name = "rusticwormhole"
-version = "0.1.0"
+version = "0.2.0"
 authors = ["Esteban Küber <esteban@kuber.com.ar>"]
-edition = "2018"
+edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-clap = "3.0.0-beta.2"
-async-std = { version = "1.6.5", features = ["attributes"] }
-tide = "0.13.0"
-serde = "*"
-surf = "2.0.0"
+clap = { version = "3.2.21", features = ["derive"] }
+tokio = { version = "1.21.1", features = ["net", "io-std", "io-util", "macros", "rt-multi-thread"] }
+warp = "0.3.2"
+hyper = { version = "0.14", features = ["full"] }
+serde = { version = "1.0.144", features = ["derive"] }
+serde_json = "1.0.85"
+http = "*"
+futures-util = "*"
 async-tls = "0.8"
